Split token into Yao::Token
